Update distance calculation to use double precision for accuracy
Change the distance property type from int to double in GeoapifyLocationService.

+++ src/Services/GeoapifyLocationService.cs
@@ -186,7 +186,7 @@ public class GeoapifyLocationService : IDistanceCalculator, IReverseGeocoder
         public string? Units { get; set; }
 
         [JsonPropertyName("distance")]
-        public int Distance { get; set; }
+        public double Distance { get; set; }
 
         [JsonPropertyName("distance_units")]
         public string? DistanceUnits { get; set; }
